+++ packages/uql-orm/src/querier/abstractQuerierPool.ts
@@ -10,7 +10,6 @@ import type {
   QueryAggMap,
   QueryAggregate,
   QueryAggregateResult,
-  QueryFindResult,
   QueryGroupMap,
   QueryOne,
   QueryOptions,
@@ -59,36 +58,24 @@ export abstract class AbstractQuerierPool<Q extends Querier, D extends AbstractD
     return context ? withContext(context, fn) : fn();
   }
 
-  findOneById<E extends object, const Q extends QueryOne<E>>(
+  findOneById<E extends object>(
     entity: Type<E>,
     id: IdValue<E>,
-    q?: Q,
+    q?: QueryOne<E>,
     opts?: QueryOptions,
-  ): Promise<QueryFindResult<E, Q> | undefined> {
+  ): Promise<E | undefined> {
     return this.withQuerier((querier) => querier.findOneById(entity, id, q, opts));
   }
 
-  findOne<E extends object, const Q extends QueryOne<E>>(
-    entity: Type<E>,
-    q: Q,
-    opts?: QueryOptions,
-  ): Promise<QueryFindResult<E, Q> | undefined> {
+  findOne<E extends object>(entity: Type<E>, q: QueryOne<E>, opts?: QueryOptions): Promise<E | undefined> {
     return this.withQuerier((querier) => querier.findOne(entity, q, opts));
   }
 
-  findMany<E extends object, const Q extends Query<E>>(
-    entity: Type<E>,
-    q: Q,
-    opts?: QueryOptions,
-  ): Promise<QueryFindResult<E, Q>[]> {
+  findMany<E extends object>(entity: Type<E>, q: Query<E>, opts?: QueryOptions): Promise<E[]> {
     return this.withQuerier((querier) => querier.findMany(entity, q, opts));
   }
 
-  findManyAndCount<E extends object, const Q extends Query<E>>(
-    entity: Type<E>,
-    q: Q,
-    opts?: QueryOptions,
-  ): Promise<[QueryFindResult<E, Q>[], number]> {
+  findManyAndCount<E extends object>(entity: Type<E>, q: Query<E>, opts?: QueryOptions): Promise<[E[], number]> {
     return this.withQuerier((querier) => querier.findManyAndCount(entity, q, opts));
   }